chore: philo.h および関連ファイルの時間データ型を「long」に更新します

+++ srcs/ft_atoll.c
@@ -6,15 +6,15 @@
 /*   By: ootsuboyoshiyuki <ootsuboyoshiyuki@stud    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/25 14:36:13 by ootsuboyosh       #+#    #+#             */
-/*   Updated: 2024/05/26 15:43:24 by ootsuboyosh      ###   ########.fr       */
+/*   Updated: 2024/05/27 17:32:32 by ootsuboyosh      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long long	ft_atoll(const char *str)
+long	ft_atoll(const char *str)
 {
-	long long	res;
+	long	res;
 	int			sign;
 
 	res = 0;
